requirements.txt
fastapi
uvicorn
pydantic
httpx
PyGithub
openai
python-dotenv

